+++ crates/masm-analysis/src/unconstrained_advice/shared.rs
@@ -6,7 +6,7 @@ pub(super) use masm_decompiler::{
     intrinsic_nonzero_arg_index, intrinsic_positional_u32_arg_range,
     intrinsic_requires_u32_precondition,
 };
-use masm_decompiler::{Intrinsic, LocalAccessKind, LoopPhi, Stmt, Var};
+use masm_decompiler::{Intrinsic, LoopPhi, Stmt, Var};
 
 use super::{domain::AdviceFact, u32_domain::U32Validity};
 pub(super) use super::{
@@ -14,6 +14,9 @@ pub(super) use super::{
     expr::{
         assign_expr_metadata, collect_expr_sink_fact, expr_is_proven_nonzero, expr_output_fact,
         expr_u32_validity, refine_if_envs,
+    },
+    local_transfer::{
+        apply_local_load_scalar, apply_local_load_word, apply_local_store, apply_local_store_word,
     },
 };
 use crate::abstract_interp::{FixpointConfig, JoinSemiLattice, iterate_to_fixpoint};
@@ -232,77 +235,6 @@ pub(super) fn apply_intrinsic_effect(
     }
 }
 
-/// Store a scalar local and preserve exact alias metadata when possible.
-pub(super) fn apply_local_store(values: &[Var], index: u32, env: &mut Env) {
-    let fact = AdviceFact::join_all(values.iter().map(|var| env.fact_for_var(var)));
-    env.set_local_fact(index, fact);
-    let validity = single_var(values)
-        .map(|var| env.u32_validity_for_var(var))
-        .unwrap_or(U32Validity::Unknown);
-    env.set_local_u32_validity(index, validity);
-    let identity = single_var(values).map(|var| env.identity_for_var(var));
-    let witness = single_var(values).and_then(|var| env.zero_test_for_var(var));
-    env.set_local_identity(index, identity);
-    env.set_local_zero_test(index, witness);
-}
-
-/// Store a local word slot-by-slot.
-pub(super) fn apply_local_store_word(
-    kind: LocalAccessKind,
-    values: &[Var],
-    index: u32,
-    env: &mut Env,
-) {
-    for (offset, value) in values.iter().enumerate() {
-        let slot = match kind {
-            // Canonical slot order follows ascending local-memory addresses.
-            LocalAccessKind::WordBe => index + (values.len().saturating_sub(1) - offset) as u32,
-            LocalAccessKind::WordLe => index + offset as u32,
-            LocalAccessKind::Element => index,
-        };
-        env.set_local_fact(slot, env.fact_for_var(value));
-        env.set_local_u32_validity(slot, env.u32_validity_for_var(value));
-        env.set_local_identity(slot, None);
-        env.set_local_zero_test(slot, None);
-    }
-}
-
-/// Load a scalar local and restore exact alias metadata when available.
-pub(super) fn apply_local_load_scalar(outputs: &[Var], index: u32, env: &mut Env) {
-    let fact = env.fact_for_local(index);
-    let identity = env.identity_for_local(index);
-    let witness = env.zero_test_for_local(index);
-    for output in outputs {
-        env.set_var_fact(output, fact.clone());
-        env.set_var_u32_validity(output, env.u32_validity_for_local(index));
-        if let Some(identity) = identity.clone() {
-            env.set_var_identity(output, identity);
-        } else {
-            env.clear_var_identity(output);
-        }
-        env.set_var_zero_test(output, witness.clone());
-    }
-}
-
-/// Load a local word slot-by-slot, preserving stack order.
-pub(super) fn apply_local_load_word(
-    kind: LocalAccessKind,
-    outputs: &[Var],
-    index: u32,
-    env: &mut Env,
-) {
-    for (offset, output) in outputs.iter().enumerate() {
-        let slot = match kind {
-            LocalAccessKind::WordBe => index + offset as u32,
-            LocalAccessKind::WordLe => index + (outputs.len().saturating_sub(1) - offset) as u32,
-            LocalAccessKind::Element => index,
-        };
-        env.set_var_fact(output, env.fact_for_local(slot));
-        env.set_var_u32_validity(output, env.u32_validity_for_local(slot));
-        env.clear_var_metadata(output);
-    }
-}
-
 /// Return the statement span for a structured statement.
 pub(super) fn stmt_span(stmt: &Stmt) -> miden_debug_types::SourceSpan {
     match stmt {
@@ -323,14 +255,6 @@ pub(super) fn stmt_span(stmt: &Stmt) -> miden_debug_types::SourceSpan {
         | Stmt::If { span, .. }
         | Stmt::While { span, .. }
         | Stmt::Return { span, .. } => *span,
-    }
-}
-
-/// Return the sole variable in the slice, if there is exactly one.
-pub(super) fn single_var(vars: &[Var]) -> Option<&Var> {
-    match vars {
-        [var] => Some(var),
-        _ => None,
     }
 }
